added JSON.stringify() to handleSubmit
I fixed my "body:" output thanks to Udacity suggestion: "You've forgotten to stringify the request you're making to the server. You've indicated in the Content-Type header of the request that you're sending a JSON object in the request body, but in fact you're sending a plain JS object. Simply use JSON.stringify and the request will be successful."